+++ tests/features/label_calibration/test_simple/test_calibrator.py
@@ -16,7 +16,7 @@ import numpy as np
 import pytest
 from sklearn.exceptions import NotFittedError
 
-from qualle.features.label_calibration.simple_label_calibration import (
+from qualle.features.label_calibration.simple import (
     SimpleLabelCalibrator,
 )
 from tests.common import DummyRegressor
