+++ tests/features/label_calibration/test_simple/test_features.py
@@ -15,7 +15,7 @@
 import numpy as np
 import pytest
 
-from qualle.features.label_calibration.simple_label_calibration import (
+from qualle.features.label_calibration.simple import (
     SimpleLabelCalibrationFeatures,
 )
 from qualle.models import LabelCalibrationData
